Add: Auth login,logout,signup && Chore: Users remove methods
