+++ src/auth/entities/invite.entity.ts
@@ -1,0 +1,32 @@
+import { Column, CreateDateColumn, DeleteDateColumn, Entity, PrimaryGeneratedColumn, UpdateDateColumn } from "typeorm";
+
+@Entity({
+  name: 'invites'
+})
+
+export class Invites {
+  @PrimaryGeneratedColumn()
+  id : number;
+
+  @Column('varchar')
+  email : string;
+
+  @Column({ type : 'varchar' , nullable : false })
+  token : string;
+
+  @Column({ type : 'timestamp' , nullable : false })
+  expires : Date;
+
+  @Column({ type : 'varchar', nullable : false })
+  status : string;
+
+  @CreateDateColumn({ type : 'timestamp', nullable : false })
+  createdAt : Date;
+
+  @UpdateDateColumn({ type : 'timestamp', nullable : false })
+  updatedAt : Date;
+
+  @DeleteDateColumn({ type : 'timestamp', nullable : true })
+  deletedAt? : Date;
+
+}
